+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # version 0.0.2
- - Properly added iai-oop to the package dependencies
+  - Properly added iai-oop to the package dependencies
+  - Added 'iaiComponent.require'
+  - Modified the error prompt for enhance debugging
 
 # version 0.0.1
  - Initial version of the component. The API is experimental but stable.